% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_match.R
\name{cr_match}
\alias{cr_match}
\title{Match free-form references to DOIs}
\usage{
cr_match(refs = NULL)
}
\arguments{
\item{refs}{character, references to be matched.}
}
\description{
Retrieve DOIs by posting free-form references to Crossref.
}
\examples{
\dontrun{
cr_match("Kleinbölting N, Huep G, Weisshaar B. Enhancing the GABI-Kat Arabidopsis thaliana T-DNA insertion mutant database by incorporating Araport11 annotation. Plant and Cell Physiology. 2017;58(1): e7.")
}
}

