% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_match.R
\name{cr_match_}
\alias{cr_match_}
\title{Post one request to match many free-form reference to DOIs}
\usage{
cr_match_(refs = NULL)
}
\arguments{
\item{refs}{character, references to be matched}
}
\description{
In general, use \link{cr_match} instead. It calls this function,
posting free-form references to Crossref to retrieve DOIS for every match.
}

