% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmatch-package.R
\docType{package}
\name{roadoi-package}
\alias{roadoi}
\alias{roadoi-package}
\title{crmmatch - Match references to DOIs using the Crossref search}
\description{
crmmatch - Match references to DOIs using the Crossref search
}
\keyword{crossref}
\keyword{package}

